+++ src/main/java/com/attica/athens/global/security/jwt/JWTUtil.java
@@ -1,14 +1,13 @@
-package com.attica.athens.global.security;
+package com.attica.athens.global.security.jwt;
 
+import com.attica.athens.global.security.userdetail.CustomUserDetails;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.Jwts.SIG;
-import jakarta.servlet.http.HttpServletRequest;
 import java.nio.charset.StandardCharsets;
 import java.util.Date;
 import javax.crypto.SecretKey;
 import javax.crypto.spec.SecretKeySpec;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.security.authentication.AuthenticationCredentialsNotFoundException;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Component;
