+++ src/main/java/com/attica/athens/global/security/refresh/domain/RefreshToken.java
@@ -1,11 +1,10 @@
-package com.attica.athens.global.security.token.domain;
+package com.attica.athens.global.security.refresh.domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import java.time.LocalDate;
 import java.util.Date;
 import lombok.AccessLevel;
 import lombok.Getter;
